requirements.txt
cmake
dlib
wheel
psycopg2
psycopg2-binary
face_recognition
numpy~=1.26.4
opencv-python~=4.10.0.84
torch~=2.6.0
PyYAML~=6.0.2
pillow~=11.1.0
torchvision~=0.21.0
matplotlib~=3.10.1
ultralytics~=8.3.82
pandas~=2.2.3
seaborn~=0.13.2
scipy~=1.15.2
requests~=2.32.3
tqdm~=4.67.1
psutil~=7.0.0
thop~=0.1.1-2209072238
ultralytics-thop~=2.0.14
face-recognition~=1.3.0
python-dotenv